PR6/requirements.txt

certifi==2026.2.25
charset-normalizer==3.4.4
grpcio==1.78.0
grpcio-tools==1.78.0
idna==3.11
mypy-protobuf==5.0.0
protobuf==6.33.5
requests==2.32.5
setuptools==82.0.0
types-protobuf==6.32.1.20260221
typing_extensions==4.15.0
urllib3==2.6.3

